Split auth to separate modules
